pyproject.toml
[tool.poetry]
name = "fastapi-blog"
version = "0.1.0"
description = "FastApi App for Web Blogging"
authors = ["Smetanin Aleksandr <smetamx@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.61.2"
uvicorn = "^0.12.2"
pydantic = "^1.7.2"
python-dotenv = "^0.15.0"
Jinja2 = "^2.11.2"
aiofiles = "^0.6.0"
alembic = "^1.4.3"
databases = {extras = ["sqlite"], version = "^0.4.0"}
markdown2 = "^2.3.10"

[tool.poetry.dev-dependencies]
pytest = "^6.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
